Expand WalletBlockTools to non-wallet test directories
Move the ignore_block_validation fixture from the wallet conftest to the
root conftest (non-autouse) so any test directory can opt in. Add autouse
wrappers in core/mempool, simulation, pools, fee_estimation, and
core/data_layer to skip expensive VDF/PoS computation in those suites.
Tests that need real consensus validation can opt out with
@pytest.mark.standard_block_tools.
